Fix global state in KillInvocation test, squash RawHandler test in Ingress test
